streamlit/requirements.txt
plotly-express
pycaret
streamlit_option_menu==0.3.2
matplotlib==3.5.1
nltk==3.7
numpy==1.21.5
pandas==1.4.2
Pillow==9.3.0
plotly==5.6.0
scikit_learn==1.1.3
streamlit==1.13.0